(chore) Install sequelize dependencies and refactor tests
