+++ package.json
@@ -33,7 +33,9 @@
     "dotenv": "^6.2.0",
     "express": "^4.16.4",
     "mocha": "^5.2.0",
-    "request": "^2.88.0"
+    "pg": "^7.8.0",
+    "request": "^2.88.0",
+    "sequelize": "^4.42.0"
   },
   "devDependencies": {
     "@babel/core": "^7.2.2",
